DBZ-6723 Allow specifying partitions for EventHub
Adds supports to detect the number of partitions in the supplied
EventHub producer, create batches for every partition, and to assign
EventData to a specific partition/batch, in a new BatchManager class.

When a batch hits its maximum size (1024KB), the batch is emitted and a
new batch is created.

When all incoming records have been processed all open/remaining batches
are emitted if they have any EventData assigned.

As creating an EventDataBatch initiates a connection over AMQP, that
adds quite a bit of overhead. By putting the batch creation behind a
proxy, we only create the real EventDataBatch when it is necessary,
saving a bit of overhead.